refactor: add all types for tweets except streaming one
