+++ v2/rest/index.ts
@@ -2,5 +2,7 @@ export * from './list';
 export * from './misc';
 export * from './space';
 export * from './stream';
-export * from './tweet';
-export * from './user';
+
+export * from './users';
+
+export * from './tweets';
