+++ v2/rest/tweets/index.ts
@@ -1,4 +1,7 @@
+export * from './HideReplies';
+export * from './Likes';
 export * from './ManageTweets';
+export * from './Retweets';
 export * from './SearchTweets';
 export * from './Timelines';
 export * from './TweetCounts';
