+++ v2/rest/users/index.ts
@@ -1,0 +1,1 @@
+export * from './UsersLookup';
